+++ src/oms_sensemaking/inference/rules/in_out_garrison.py
@@ -63,12 +63,6 @@ class InOrOutOfGarrison(Sensemaker):
         garrison_lat_lon = garrison_data.garrison_lat_lon
         activities = garrison_data.activities
 
-        # Check if observation has already been processed after getting coords
-        # We could fail early and never call has_action_already_ran
-        # if 'object_and_garrison_coords' conditional does not pass
-        if self.has_action_already_ran(obs.id, activities):
-            return []
-
         in_garrison_check = in_garrison(object_lat_lon, garrison_lat_lon)
         garrison_buffer_points = generate_circle_points_geographical(
             garrison_lat_lon[0], garrison_lat_lon[1], SETTINGS.garrison_distance_kilometers
@@ -164,19 +158,3 @@ class InOrOutOfGarrison(Sensemaker):
             endTime=observation.endTime,
         )
         self.oms_crud_tool.create_activity(garrison_activity)
-
-    def has_action_already_ran(self, observation_id, activities):
-        """
-        Determine if an in/out of garrison activity pointing to the inputted observation has already been created
-        """
-
-        return any(
-            act.name
-            in {
-                SETTINGS.inference_in_garrison_activity_name,
-                SETTINGS.inference_out_of_garrison_activity_name,
-            }
-            and act.observationIds
-            and observation_id in act.observationIds
-            for act in activities
-        )